优化opencv构建脚本 (#139) - 编译所有第三方组件，减少系统依赖 - 启用 IPP、LAPACK、EIGEN

+++ cpp/tools/linux_build_opencv.sh
@@ -4,17 +4,27 @@
 source_dir=$(pwd)  # 当前工作目录
 
 # 获取 OpenCV 源码目录（可相对工作目录）
+if [ $# -lt 1 ]; then
+    echo "Usage: build_opencv.sh /path/to/opencv"
+    exit 1
+fi
 opencv_dir=$1
-
 opencv_dir=$(realpath "$opencv_dir")
 cd "$opencv_dir"
 
 build_dir="${opencv_dir}/build"  # 构建目录
 release_dir="${source_dir}/opencv-release"  # 生成目录
 
-rm -rf ${build_dir}
-rm -rf ${release_dir}
+if [ -d ${build_dir} ]; then
+    rm -rf ${build_dir}
+    echo "Delete: ${build_dir}"
+fi
+if [ -d ${release_dir} ]; then
+    rm -rf ${release_dir}
+    echo "Delete: ${release_dir}"
+fi
 mkdir ${build_dir}
+echo "Create: ${build_dir}"
 cd ${build_dir}
 
 cmake ../ \
@@ -23,18 +33,19 @@ cmake ../ \
     -DBUILD_LIST=core,imgcodecs,imgproc \
     -DBUILD_SHARED_LIBS=ON \
     -DBUILD_opencv_world=OFF \
-    -DWITH_IPP=OFF \
-    -DBUILD_IPP_IW=OFF \
-    -DWITH_LAPACK=OFF \
-    -DWITH_EIGEN=OFF \
+    -DOPENCV_FORCE_3RDPARTY_BUILD=ON \
     -DWITH_ZLIB=ON \
-    -DBUILD_ZLIB=ON \
-    -DWITH_JPEG=ON \
-    -DBUILD_JPEG=ON \
-    -DWITH_PNG=ON \
-    -DBUILD_PNG=ON \
     -DWITH_TIFF=ON \
-    -DBUILD_TIFF=ON
+    -DWITH_OPENJPEG=ON \
+    -DWITH_JASPER=ON \
+    -DWITH_JPEG=ON \
+    -DWITH_PNG=ON \
+    -DWITH_OPENEXR=ON \
+    -DWITH_WEBP=ON \
+    -DWITH_IPP=ON \
+    -DWITH_LAPACK=ON \
+    -DWITH_EIGEN=ON
 
 make -j
 make install
+
